pyproject.toml
[project]
name = "dps-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.7.0",
    "fastapi>=0.104.0",
    "google-generativeai>=0.3.0",
    "openai>=1.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pymongo>=4.6.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "uvicorn>=0.24.0",
]
